Voxmate_ An Intelligent voice and vision assistant/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-noise-cancellation
openai
deepgram-sdk
duckduckgo-search
langchain_community
requests
python-dotenv
